requirements.txt
pytest>=7.0
selenium>=4.10
webdriver-manager>=4.0
allure-pytest>=2.13.3
requests>=2.28

pytest-html
pytest-xdist

